requirements.txt
python-telegram-bot==21.9
requests==2.32.3